+++ src/renderer/components/dialogs/Settings-Advanced.tsx
@@ -1,6 +1,7 @@
+import { H5 } from '@blueprintjs/core'
 import React from 'react'
 import { SettingsStoreState } from '../../stores/settings'
-import { RenderDeltaSwitch2Type } from './Settings'
+import { RenderDeltaSwitch2Type, RenderDTSettingSwitchType } from './Settings'
 import SettingsEncryption from './Settings-Encryption'
 import SettingsImapFolderHandling from './Settings-ImapFolderHandling'
 import SettingsManageKeys from './Settings-ManageKeys'
@@ -8,9 +9,11 @@ import SettingsManageKeys from './Settings-ManageKeys'
 export function SettingsAdvanced({
   settingsStore,
   renderDeltaSwitch2,
+  renderDTSettingSwitch,
 }: {
   settingsStore: SettingsStoreState
   renderDeltaSwitch2: RenderDeltaSwitch2Type
+  renderDTSettingSwitch: RenderDTSettingSwitchType
 }) {
   return (
     <>
@@ -24,6 +27,38 @@ export function SettingsAdvanced({
       <br />
       <br />
       <SettingsManageKeys />
+      <br />
+      <br />
+      <SettingsTrayIcon
+        settingsStore={settingsStore}
+        renderDTSettingSwitch={renderDTSettingSwitch}
+      />
+    </>
+  )
+}
+
+function SettingsTrayIcon({
+  settingsStore,
+  renderDTSettingSwitch,
+}: {
+  settingsStore: SettingsStoreState
+  renderDTSettingSwitch: RenderDTSettingSwitchType
+}) {
+  const tx = window.static_translate
+  return (
+    <>
+      <H5>{tx('pref_system_integration_menu_title')}</H5>
+      {renderDTSettingSwitch({
+        key: 'minimizeToTray',
+        label: tx('pref_show_tray_icon'),
+        disabled: settingsStore.rc.minimized,
+        disabledValue: settingsStore.rc.minimized,
+      })}
+      {settingsStore.rc.minimized && (
+        <div className='bp4-callout'>
+          {tx('explain_desktop_minimized_disabled_tray_pref')}
+        </div>
+      )}
     </>
   )
 }